Tests/LibWeb/Text/expected/select.txt

1. "one"
2. 0
3. "two"
4. 3
5. "three"
